refactor: 메서드와 객체의 네이밍 및 패키지 구조 변경
