+++ src/main/java/lotto/domain/rule/Rank.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.rule;
 
 import java.util.Arrays;
 import java.util.Optional;
