+++ src/main/java/lotto/domain/vo/lotto/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.vo.lotto;
 
 import java.util.HashSet;
 import java.util.List;
