+++ src/main/java/lotto/domain/vo/winning/BonusNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.vo.winning;
 
 import java.util.List;
 import lotto.exception.LottoError;
