+++ src/main/java/lotto/domain/vo/winning/WinningNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.vo.winning;
 
 import java.util.HashSet;
 import java.util.List;
