+++ src/main/java/lotto/dto/StatisticDto.java
@@ -1,7 +1,7 @@
 package lotto.dto;
 
 import java.util.EnumMap;
-import lotto.domain.Rank;
+import lotto.domain.rule.Rank;
 
 public record StatisticDto(EnumMap<Rank, Integer> statisticSummary, float profitRate) {
 }
