+++ src/main/java/lotto/service/LotteryService.java
@@ -4,23 +4,23 @@ import java.util.ArrayList;
 import java.util.EnumMap;
 import java.util.List;
 import java.util.Optional;
-import lotto.domain.BonusNumber;
-import lotto.domain.Lotto;
-import lotto.domain.LottoCalculator;
-import lotto.domain.Rank;
+import lotto.domain.LottoMatcher;
 import lotto.domain.StatisticCalculator;
-import lotto.domain.WinningNumber;
+import lotto.domain.rule.Rank;
 import lotto.domain.vo.PurchaseAmount;
+import lotto.domain.vo.lotto.Lotto;
+import lotto.domain.vo.winning.BonusNumber;
+import lotto.domain.vo.winning.WinningNumber;
 import lotto.dto.StatisticDto;
 import lotto.dto.mapper.DtoMapper;
 
 public class LotteryService {
 
-    private final LottoCalculator lottoCalculator;
+    private final LottoMatcher lottoMatcher;
     private final StatisticCalculator statisticCalculator;
 
-    public LotteryService(LottoCalculator lottoCalculator, StatisticCalculator statisticCalculator) {
-        this.lottoCalculator = lottoCalculator;
+    public LotteryService(LottoMatcher lottoMatcher, StatisticCalculator statisticCalculator) {
+        this.lottoMatcher = lottoMatcher;
         this.statisticCalculator = statisticCalculator;
     }
 
@@ -30,9 +30,7 @@ public class LotteryService {
 
         lottos.forEach(lotto -> {
             Optional<Rank> rank = findRankWith(lotto, winningNumbers, bonus);
-            if (rank.isPresent()) {
-                ranks.add(rank.get());
-            }
+            rank.ifPresent(ranks::add);
         });
 
         EnumMap<Rank, Integer> statisticSummary = statisticCalculator.summarizeCountOf(ranks);
@@ -43,8 +41,8 @@ public class LotteryService {
 
     public Optional<Rank> findRankWith(Lotto lotto, WinningNumber winningNumber,
                                        BonusNumber bonusNumber) {
-        int matchingCount = lottoCalculator.findMatchingCount(lotto, winningNumber);
-        boolean matchBonus = lottoCalculator.matchValuesOf(lotto, bonusNumber);
+        int matchingCount = lottoMatcher.findMatchingCount(lotto, winningNumber);
+        boolean matchBonus = lottoMatcher.containValueOf(lotto, bonusNumber);
 
         return Rank.findBy(matchingCount, matchBonus);
     }
