+++ src/test/java/lotto/domain/rule/RandomPickerTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.rule;
 
 import java.util.HashSet;
 import java.util.List;
