+++ src/test/java/lotto/service/LotteryServiceTest.java
@@ -2,12 +2,12 @@ package lotto.service;
 
 import java.util.List;
 import java.util.Optional;
-import lotto.domain.BonusNumber;
-import lotto.domain.Lotto;
-import lotto.domain.LottoCalculator;
-import lotto.domain.Rank;
+import lotto.domain.LottoMatcher;
 import lotto.domain.StatisticCalculator;
-import lotto.domain.WinningNumber;
+import lotto.domain.rule.Rank;
+import lotto.domain.vo.lotto.Lotto;
+import lotto.domain.vo.winning.BonusNumber;
+import lotto.domain.vo.winning.WinningNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -20,9 +20,9 @@ public class LotteryServiceTest {
     void 로또와_당첨_번호_보너스_번호에_일치하는_등수를_찾아_반환한다(List<Integer> lottoNumbers, List<Integer> winningNumbers, int numberValue,
                                             Rank rank) {
         //given
-        LottoCalculator calculator = new LottoCalculator();
+        LottoMatcher matcher = new LottoMatcher();
         StatisticCalculator statistic = new StatisticCalculator();
-        LotteryService service = new LotteryService(calculator, statistic);
+        LotteryService service = new LotteryService(matcher, statistic);
 
         //when
         Lotto lotto = new Lotto(lottoNumbers);
